+++ src/store/index.ts
@@ -19,7 +19,7 @@ export interface StoreOverlayPilot extends StoreOverlayDefault {
     type: 'pilot';
     data: {
         pilot: VatsimExtendedPilot
-        stats?: VatsimMemberStats
+        stats?: Partial<VatsimMemberStats>
     };
 }
 
@@ -30,6 +30,7 @@ export const useStore = defineStore('index', {
         theme: 'default' as ThemesList,
         extent: [0, 0, 0, 0] as Extent,
         zoom: 0,
+        datetime: Date.now(),
         moving: false,
         openOverlayId: null as string | null,
         openPilotOverlay: false,
@@ -45,18 +46,43 @@ export const useStore = defineStore('index', {
         },
     },
     actions: {
-        addOverlay(overlay: Pick<StoreOverlay, 'data' | 'type' | 'maxHeight' | 'sticky'>) {
+        addOverlay<O extends StoreOverlay = StoreOverlay>(overlay: Pick<O, 'data' | 'type' | 'maxHeight' | 'sticky'>) {
             const id = (this.overlays[this.overlays.length - 1]?.id ?? 0) + 1;
             for (const overlay of this.overlays.filter(x => typeof x.position === 'number')) {
                 (overlay.position as number)++;
             }
 
-            this.overlays.push({
+            const newOverlay: StoreOverlay = {
                 id,
                 position: 0,
                 collapsed: false,
                 ...overlay,
+            };
+
+            this.overlays.push(newOverlay);
+            return this.overlays.find(x => x.id === id)!;
+        },
+        async addPilotOverlay(cid: string) {
+            const existingOverlay = this.overlays.find(x => x.type === 'pilot');
+
+            if (existingOverlay?.data.pilot.cid.toString() === cid) return;
+            if (existingOverlay && !existingOverlay?.sticky) this.overlays = this.overlays.filter(x => x.id !== existingOverlay.id);
+
+            const pilot = await $fetch(`/data/vatsim/pilot/${ cid }`);
+
+            const overlay = this.addOverlay({
+                data: {
+                    pilot,
+                    stats: {
+                        pilot: 0,
+                    },
+                },
+                type: 'pilot',
+                maxHeight: 45,
+                sticky: cid === this.user?.cid,
             });
+
+            overlay.data.stats = await $fetch<VatsimMemberStats>(`/data/vatsim/pilot/${ cid }/stats`);
         },
     },
 });