+++ src/utils/index.ts
@@ -1,3 +1,11 @@
+import { useStore } from '~/store';
+
 export function sleep(ms: number) {
     return new Promise(resolve => setTimeout(resolve, ms));
+}
+
+export function getHoursAndMinutes(date: number) {
+    const diff = Math.abs(useStore().datetime - date) / (1000 * 60);
+
+    return `${ (`0${ Math.floor(diff / 60) }`).slice(-2) }:${ (`0${ Math.floor(diff % 60) }`).slice(-2) }`;
 }